Transaction table views set up

+++ project 2/API/Responses/TransactionResponse.swift
@@ -11,7 +11,7 @@ struct TransactionResponse: Codable {
     let id: String
     let senderId: String
     let receiverId: String
-    let amount: Decimal
+    let amount: Double
     let currency: String
     let createdOn: Int
     let reference: String
